requirements.txt
playwright~=1.36.0
aiohttp~=3.8.4
PyYAML~=6.0.1
setuptools~=65.5.0